Move helper functions out into separate file
