+++ viam-cartographer.go
@@ -9,8 +9,6 @@ import (
 	"image"
 	"image/jpeg"
 	"io"
-	"os"
-	"path/filepath"
 	"strconv"
 	"strings"
 	"sync"
@@ -66,16 +64,6 @@ type SubAlgo string
 // Dim2d runs cartographer with a 2D LIDAR only.
 const Dim2d SubAlgo = "2d"
 
-// SetCameraValidationMaxTimeoutSecForTesting sets cameraValidationMaxTimeoutSec for testing.
-func SetCameraValidationMaxTimeoutSecForTesting(val int) {
-	cameraValidationMaxTimeoutSec = val
-}
-
-// SetDialMaxTimeoutSecForTesting sets dialMaxTimeoutSec for testing.
-func SetDialMaxTimeoutSecForTesting(val int) {
-	dialMaxTimeoutSec = val
-}
-
 func init() {
 	registry.RegisterService(slam.Subtype, Model, registry.Service{
 		Constructor: func(ctx context.Context, deps registry.Dependencies, config config.Service, logger golog.Logger) (interface{}, error) {
@@ -97,46 +85,104 @@ func init() {
 		}, &slamConfig.AttrConfig{})
 }
 
-// runtimeServiceValidation ensures the service's data processing and saving is valid for the subAlgo and
-// cameras given.
-func runtimeServiceValidation(
+// New returns a new slam service for the given robot.
+func New(
 	ctx context.Context,
-	cams []camera.Camera,
-	cartoSvc *cartographerService,
-) error {
-	if !cartoSvc.useLiveData {
-		return nil
+	deps registry.Dependencies,
+	config config.Service,
+	logger golog.Logger,
+	bufferSLAMProcessLogs bool,
+	executableName string,
+) (slam.Service, error) {
+	ctx, span := trace.StartSpan(ctx, "viamcartographer::slamService::New")
+	defer span.End()
+
+	svcConfig, ok := config.ConvertedAttributes.(*slamConfig.AttrConfig)
+	if !ok {
+		return nil, rdkutils.NewUnexpectedTypeError(svcConfig, config.ConvertedAttributes)
 	}
 
-	var err error
-	var path string
-	paths := make([]string, 0, 1)
-	startTime := time.Now()
-
-	for {
-		path, err = cartoSvc.getAndSaveData(ctx, cams)
-		paths = append(paths, path)
-
-		if err == nil {
-			break
-		}
-
-		// This takes about 5 seconds, so the timeout should be sufficient.
-		if time.Since(startTime) >= time.Duration(cameraValidationMaxTimeoutSec)*time.Second {
-			return errors.Wrap(err, "error getting data from sensor")
-		}
-		if !goutils.SelectContextOrWait(ctx, cameraValidationIntervalSec*time.Second) {
-			return ctx.Err()
-		}
+	primarySensorName, cams, err := configureCameras(svcConfig, deps)
+	if err != nil {
+		return nil, errors.Wrap(err, "configuring camera error")
 	}
 
-	for _, path := range paths {
-		if err := os.RemoveAll(path); err != nil {
-			return errors.Wrap(err, "error removing generated file during validation")
-		}
+	subAlgo := SubAlgo(svcConfig.ConfigParams["mode"])
+	if subAlgo != Dim2d {
+		return nil, errors.Errorf("%v does not have a 'mode: %v'",
+			string(config.Model.Name), svcConfig.ConfigParams["mode"])
 	}
 
-	return nil
+	// Set up the data directories
+	err = slamConfig.SetupDirectories(svcConfig.DataDirectory, logger)
+	if err != nil {
+		return nil, err
+	}
+
+	port, dataRateMsec, mapRateSec, useLiveData, deleteProcessedData, err := slamConfig.GetOptionalParameters(
+		svcConfig,
+		localhost0,
+		defaultDataRateMsec,
+		defaultMapRateSec,
+		logger,
+	)
+	if useLiveData {
+		logger.Debug("Running in live mode")
+	} else {
+		logger.Debug("Running in offline mode")
+	}
+	if err != nil {
+		return nil, err
+	}
+	cancelCtx, cancelFunc := context.WithCancel(ctx)
+
+	// SLAM Service Object
+	cartoSvc := &cartographerService{
+		primarySensorName:     primarySensorName,
+		executableName:        executableName,
+		subAlgo:               subAlgo,
+		slamProcess:           pexec.NewProcessManager(logger),
+		configParams:          svcConfig.ConfigParams,
+		dataDirectory:         svcConfig.DataDirectory,
+		useLiveData:           useLiveData,
+		deleteProcessedData:   deleteProcessedData,
+		port:                  port,
+		dataRateMs:            dataRateMsec,
+		mapRateSec:            mapRateSec,
+		cancelFunc:            cancelFunc,
+		logger:                logger,
+		bufferSLAMProcessLogs: bufferSLAMProcessLogs,
+		dev:                   svcConfig.Dev,
+	}
+
+	var success bool
+	defer func() {
+		if !success {
+			if err := cartoSvc.Close(); err != nil {
+				logger.Errorw("error closing out after error", "error", err)
+			}
+		}
+	}()
+
+	if err := runtimeServiceValidation(cancelCtx, cams, cartoSvc); err != nil {
+		return nil, errors.Wrap(err, "runtime slam service error")
+	}
+
+	cartoSvc.StartDataProcess(cancelCtx, cams, nil)
+
+	if err := cartoSvc.StartSLAMProcess(ctx); err != nil {
+		return nil, errors.Wrap(err, "error with slam service slam process")
+	}
+
+	client, clientClose, err := slamConfig.SetupGRPCConnection(ctx, cartoSvc.port, dialMaxTimeoutSec, logger)
+	if err != nil {
+		return nil, errors.Wrap(err, "error with initial grpc client to slam algorithm")
+	}
+	cartoSvc.clientAlgo = client
+	cartoSvc.clientAlgoClose = clientClose
+
+	success = true
+	return cartoSvc, nil
 }
 
 // cartographerService is the structure of the slam service.
@@ -168,30 +214,6 @@ type cartographerService struct {
 	slamProcessLogReader         io.ReadCloser
 	slamProcessLogWriter         io.WriteCloser
 	slamProcessBufferedLogReader bufio.Reader
-}
-
-// configureCameras will check the config to see if any cameras are desired and if so, grab the cameras from
-// the robot. We assume there is at most one lidar camera.
-func configureCameras(
-	svcConfig *slamConfig.AttrConfig,
-	deps registry.Dependencies,
-) (string, []camera.Camera, error) {
-	if len(svcConfig.Sensors) == 0 {
-		return "", nil, nil
-	}
-	if len(svcConfig.Sensors) != 1 {
-		return "", nil, errors.Errorf("'sensors' must contain only one lidar camera, but is 'sensors: [%v]'",
-			strings.Join(svcConfig.Sensors, ", "))
-	}
-	cams := make([]camera.Camera, 0, len(svcConfig.Sensors))
-	// The first camera is expected to be LIDAR.
-	primarySensorName := svcConfig.Sensors[0]
-	cam, err := camera.FromDependencies(deps, primarySensorName)
-	if err != nil {
-		return "", nil, errors.Wrapf(err, "error getting camera %v for slam service", primarySensorName)
-	}
-	cams = append(cams, cam)
-	return primarySensorName, cams, nil
 }
 
 // Position forwards the request for positional data to the slam library's gRPC service. Once a response is received,
@@ -421,105 +443,6 @@ func (cartoSvc *cartographerService) GetInternalStateStream(ctx context.Context,
 	return grpchelper.GetInternalStateStreamCallback(ctx, name, cartoSvc.clientAlgo)
 }
 
-// New returns a new slam service for the given robot.
-func New(
-	ctx context.Context,
-	deps registry.Dependencies,
-	config config.Service,
-	logger golog.Logger,
-	bufferSLAMProcessLogs bool,
-	executableName string,
-) (slam.Service, error) {
-	ctx, span := trace.StartSpan(ctx, "viamcartographer::slamService::New")
-	defer span.End()
-
-	svcConfig, ok := config.ConvertedAttributes.(*slamConfig.AttrConfig)
-	if !ok {
-		return nil, rdkutils.NewUnexpectedTypeError(svcConfig, config.ConvertedAttributes)
-	}
-
-	primarySensorName, cams, err := configureCameras(svcConfig, deps)
-	if err != nil {
-		return nil, errors.Wrap(err, "configuring camera error")
-	}
-
-	subAlgo := SubAlgo(svcConfig.ConfigParams["mode"])
-	if subAlgo != Dim2d {
-		return nil, errors.Errorf("%v does not have a 'mode: %v'",
-			string(config.Model.Name), svcConfig.ConfigParams["mode"])
-	}
-
-	err = slamConfig.SetupDirectories(svcConfig.DataDirectory, logger)
-	if err != nil {
-		return nil, err
-	}
-
-	port, dataRateMsec, mapRateSec, useLiveData, deleteProcessedData, err := slamConfig.GetOptionalParameters(
-		svcConfig,
-		localhost0,
-		defaultDataRateMsec,
-		defaultMapRateSec,
-		logger,
-	)
-	if useLiveData {
-		logger.Debug("Running in live mode")
-	} else {
-		logger.Debug("Running in offline mode")
-	}
-	if err != nil {
-		return nil, err
-	}
-	cancelCtx, cancelFunc := context.WithCancel(ctx)
-
-	// SLAM Service Object
-	cartoSvc := &cartographerService{
-		primarySensorName:     primarySensorName,
-		executableName:        executableName,
-		subAlgo:               subAlgo,
-		slamProcess:           pexec.NewProcessManager(logger),
-		configParams:          svcConfig.ConfigParams,
-		dataDirectory:         svcConfig.DataDirectory,
-		useLiveData:           useLiveData,
-		deleteProcessedData:   deleteProcessedData,
-		port:                  port,
-		dataRateMs:            dataRateMsec,
-		mapRateSec:            mapRateSec,
-		cancelFunc:            cancelFunc,
-		logger:                logger,
-		bufferSLAMProcessLogs: bufferSLAMProcessLogs,
-		dev:                   svcConfig.Dev,
-	}
-
-	var success bool
-	defer func() {
-		if !success {
-			if err := cartoSvc.Close(); err != nil {
-				logger.Errorw("error closing out after error", "error", err)
-			}
-		}
-	}()
-
-	if err := runtimeServiceValidation(cancelCtx, cams, cartoSvc); err != nil {
-		return nil, errors.Wrap(err, "runtime slam service error")
-	}
-
-	cartoSvc.StartDataProcess(cancelCtx, cams, nil)
-
-	if err := cartoSvc.StartSLAMProcess(ctx); err != nil {
-		return nil, errors.Wrap(err, "error with slam service slam process")
-	}
-
-	client, clientClose, err := slamConfig.SetupGRPCConnection(ctx, cartoSvc.port, dialMaxTimeoutSec, logger)
-	if err != nil {
-		return nil, errors.Wrap(err, "error with initial grpc client to slam algorithm")
-	}
-	cartoSvc.clientAlgo = client
-	cartoSvc.clientAlgoClose = clientClose
-
-	success = true
-	return cartoSvc, nil
-}
-
 // Close out of all slam related processes.
 func (cartoSvc *cartographerService) Close() error {
 	defer func() {
@@ -735,11 +658,4 @@ func (cartoSvc *cartographerService) getAndSaveData(ctx context.Context, cams []
 	fileType := ".pcd"
 	filename := createTimestampFilename(cartoSvc.dataDirectory, cartoSvc.primarySensorName, fileType)
 	return filename, dataprocess.WritePCDToFile(pointcloud, filename)
-}
-
-// Creates a file for camera data with the specified sensor name and timestamp written into the filename.
-func createTimestampFilename(dataDirectory, primarySensorName, fileType string) string {
-	timeStamp := time.Now()
-	dataDir := filepath.Join(dataDirectory, "data")
-	return dataprocess.CreateTimestampFilename(dataDir, primarySensorName, fileType, timeStamp)
 }
